# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Whatsapp Clone API Documentation",
        "contact": {
            "email": "mail@satrio.dev"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/chatrooms/{chatroom}/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get messages in a chatroom",
                "description": "Fetches all messages in a specific chatroom with user information.",
                "operationId": "1f4fffcf2edc151fba95b4864a2a7bf5",
                "parameters": [
                    {
                        "name": "chatroom",
                        "in": "path",
                        "description": "The ID of the chatroom",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "13662cf0-ead2-48b7-a056-ded77823fe9c"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of messages in the chatroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "text": {
                                                "type": "string",
                                                "example": "Hello!"
                                            },
                                            "attachment_path": {
                                                "type": "string",
                                                "example": "storage/attachment.png"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-11-06T12:34:56"
                                            },
                                            "user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chatroom not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chatroom not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/chatrooms/{chatroomId}/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send a message to a chatroom",
                "description": "Sends a message to a chatroom, optionally with an attachment (image/video).",
                "operationId": "d06cf596d94c72de0aa5f15bec7f3914",
                "parameters": [
                    {
                        "name": "chatroomId",
                        "in": "path",
                        "description": "The ID of the chatroom",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "13662cf0-ead2-48b7-a056-ded77823fe9c"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Hello everyone!"
                                    },
                                    "attachment": {
                                        "description": "Optional file attachment (image/video)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Message queued for processing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Message queued for processing"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (either message or attachment is required, or invalid file type)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Message or attachment is required"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}